Move AlcesUtils to be part of Underware Gem's public API
Although only for use in tests, AlcesUtils will still be useful to other
code using Underware as a Gem (in particular Metalware), as it allows
the behaviour of Underware's namespaces to be mocked out for
easier/clearer testing.
