+++ spec/cache/asset_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'alces_utils'
+require 'underware/spec/alces_utils'
 require 'underware/cache/asset'
 
 RSpec.describe Underware::Cache::Asset do
