+++ spec/commands/asset/unlink_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'filesystem'
-require 'alces_utils'
+require 'underware/spec/alces_utils'
 
 RSpec.describe Underware::Commands::Asset::Unlink do
   include AlcesUtils
