+++ spec/dependency_spec.rb
@@ -30,7 +30,7 @@ require 'ostruct'
 require 'spec_helper'
 require 'fileutils'
 require 'filesystem'
-require 'alces_utils'
+require 'underware/spec/alces_utils'
 
 RSpec.describe Underware::Dependency do
   include AlcesUtils
