+++ spec/group_cache_spec.rb
@@ -24,7 +24,7 @@
 
 require 'underware/group_cache'
 require 'filesystem'
-require 'alces_utils'
+require 'underware/spec/alces_utils'
 
 RSpec.describe Underware::GroupCache do
   include AlcesUtils
