+++ spec/templating/nil_detection_wrapper_spec.rb
@@ -4,7 +4,7 @@
 require 'underware/templating/nil_detection_wrapper'
 require 'recursive_open_struct'
 require 'underware/underware_log'
-require 'alces_utils'
+require 'underware/spec/alces_utils'
 
 RSpec.describe Underware::Templating::NilDetectionWrapper do
   AlcesUtils.start self
