+++ spec/validation/answer_spec.rb
@@ -24,7 +24,7 @@
 require 'underware/validation/answer'
 require 'underware/data'
 require 'filesystem'
-require 'alces_utils'
+require 'underware/spec/alces_utils'
 
 RSpec.describe Underware::Validation::Answer do
   include AlcesUtils
